+++ tests/encodings.rs
@@ -37,7 +37,7 @@ fn test_koi8_r_encoding() {
     loop {
         match r.read_event_into(&mut buf) {
             Ok(Text(e)) => {
-                e.unescape().unwrap();
+                e.decode().unwrap();
             }
             Ok(Eof) => break,
             _ => (),